\name{bear_data_all}
\alias{bear_data_all}
\docType{data}
\title{
Bear Data All
}
\description{
This dataset includes some samples with the location, date, index, gender and locus of bears. 
}
\usage{data("bear_data_all")}
\format{
  A csv file with meta data and locus data.
}
\examples{
data(bear_data_all)
}
\keyword{datasets}
