\name{bear_data_part1}
\alias{bear_data_part1}
\docType{data}
\title{
Bear Data Part 1
}
\description{
This dataset includes some samples with the location, date, index, gender and locus of bears. 
This is a cutdown dataset, with the complementary part2 for the user to test to merge multiple sets of data.
}
\usage{data("bear_data_part1")}
\format{
  A csv file with meta data and locus data.
}
\examples{
data(bear_data_all)
}
\keyword{datasets}
