% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{calculate_new_data_distances}
\alias{calculate_new_data_distances}
\title{Calculate the distances from the new data to the existing data.}
\usage{
calculate_new_data_distances(
  new_data,
  data,
  distance_function,
  progress_object = NULL,
  treshold = NULL
)
}
\arguments{
\item{new_data}{A new_data object.}

\item{data}{A data object.}

\item{distance_function}{The function to be used for the distance. "dist_" functions are given by the package.}

\item{progress_object}{A optional shiny progress object to show progress to user}

\item{threshold}{A threshold used by some of the distance functions}
}
\value{
A multi-dimentional list with the distance from each of the new data to the data.
}
\description{
Calculate the distances from the new data to the existing data.
}
\examples{
\dontrun{
# OBS Apply the result of this function to the distance of the new_data, as this is required
# for other functions to work.

new_data$distances <- calculate_new_data_distances(new_data = new_data, data = data,
    distance_function = dist_euclidian)
}
}
