% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{combine_multilocus}
\alias{combine_multilocus}
\title{Combine A Vector Of Locuses Into A String}
\usage{
combine_multilocus(locus)
}
\arguments{
\item{locus}{A vector with the locuses.}
}
\value{
A string with all of the locuses combined.
}
\description{
Combine A Vector Of Locuses Into A String
}
\examples{
\dontrun{
print(combine_multilocus(c(182, 180, NA, 96)))
[1] "182 180 000 096"
}
}
