% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{create_new_data}
\alias{create_new_data}
\title{Create a single new data point from given data}
\usage{
create_new_data(
  index,
  multilocus,
  meta,
  na_strings = c("NA", "-99", "000", "0")
)
}
\arguments{
\item{index}{The index of this datapoint.}

\item{multilocus}{A named vector with the multilocus of this datapoint.}

\item{meta}{A named vector with all metadata for this new datapoint.}

\item{na_strings}{A string with the values that are to be interpeted as NA values.}
}
\value{
A new_data object, similar to the data object but missing values in the "individ" column.
}
\description{
Create a single new data point from given data
}
\examples{
\dontrun{
locus_data <- c("110", "112", "143", "145", "123", "127", "164", "170", "150", "150", "230",
    "248", "184", "186", "128", "132")
names(locus_data) <- c("MU09 - 1", "MU09 - 2", "MU10 - 1", "MU10 - 2", "MU05 - 1", "MU05 - 2",
    "MU23 - 1", "MU23 - 2", "MU51 - 1", "MU51 - 2", "MU59 - 1", "MU59 - 2", "G10L - 1",
    "G10L - 2", "MU50 - 1", "MU50 - 2")
new_data <- create_new_data(index = "SEP123", multilocus = locus_data, meta =
    c(date = "2020-06-29", north = "7096503", east = "644381", gender = "Hane"))
}
}
