% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{create_new_data_batch}
\alias{create_new_data_batch}
\title{Load A File Containing New Data}
\usage{
create_new_data_batch(
  file_path,
  index_column,
  locus_columns,
  meta_columns,
  na_strings = c("NA", "-99", "000"),
  sheet = 1
)
}
\arguments{
\item{file_path}{The path to the file to be loaded.}

\item{index_column}{The name of the column containing the index.}

\item{locus_columns}{A vector with the names of the columns that contains the locus data.}

\item{meta_columns}{A vector with the names of the columns that contains the meta data.}

\item{na_strings}{A vector with string that are to be interpeted as NA values.}

\item{sheet}{If a .ods or excel file is loaded, this is the sheet that will be loaded.}
}
\value{
A new_data object, similar to a data object but missing the "individ" column.
}
\description{
This is a wrapper function for load_data function.
}
\examples{
\dontrun{
See workflow.R
}
}
