% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{create_search_data}
\alias{create_search_data}
\title{Load Data and Group Into Individuals
TODO: Keep the multimatch data}
\usage{
create_search_data(
  file_path,
  index_column,
  additional_data_columns,
  locus_columns
)
}
\arguments{
\item{file_path}{A path to the csv file}

\item{index_column}{The name or index to the colmn containing the indexes for the data}

\item{additional_data_columns}{A named list with the names or indexes to the date, north, east, gender and preset_ind colmns.}

\item{locus_columns}{A vector with the names or indexes to the columns containing the locus data}
}
\value{
A list with $index, $multilocus, and $individ_id.
}
\description{
Load Data and Group Into Individuals
TODO: Keep the multimatch data
}
\examples{
\dontrun{
search_data <- create_search_data("data.csv")
}
}
