% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{dist_euclidean_num_mismatches}
\alias{dist_euclidean_num_mismatches}
\title{A distance function which uses euclidean distance, and returns the negative value if the number of mismatches is smaller than a threshold}
\usage{
dist_euclidean_num_mismatches(multilocus1, multilocus2, threshold)
}
\arguments{
\item{multilocus1}{A vector with the locuses of the first object to be compared.}

\item{multilocus2}{A vector with the locuses of the second object to be compared.}

\item{threshold}{The num-mismatch threshold where the distance that gets returned is negative}
}
\value{
The euclidean distance between locuses, negative if the number of mismatches is lower than the threshold. Missing data are counted as matches.
}
\description{
A distance function which uses euclidean distance, and returns the negative value if the number of mismatches is smaller than a threshold
}
\examples{
\dontrun{
l1 <- c(182, 180, 178, 176)
l2 <- c(178, 180, 178, 178)
dis <- dist_euclidean_num_mismatches(l1, l2, 2)
}
}
