% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{dist_manhattan}
\alias{dist_manhattan}
\title{An Manhattan Distance function}
\usage{
dist_manhattan(multilocus1, multilocus2, threshold)
}
\arguments{
\item{multilocus1}{A vector with the locuses of the first object to be compared.}

\item{multilocus2}{A vector with the locuses of the second object to be compared.}

\item{threshold}{Unused variable}
}
\value{
The manhattan distance between the locuses.
}
\description{
An Manhattan Distance function
}
\examples{
\dontrun{
l1 <- c(182, 180, 178, 176)
l2 <- c(178, 180, 178, 178)
dis <- dist_euclidean(l1, l2)
}
}
