% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{dist_maximum}
\alias{dist_maximum}
\title{An Maximum Distance function}
\usage{
dist_maximum(multilocus1, multilocus2)
}
\arguments{
\item{multilocus1}{A vector with the locuses of the first object to be compared.}

\item{multilocus2}{A vector with the locuses of the second object to be compared.}
}
\value{
The distance between the specific locuses that had the maximum distance.
}
\description{
An Maximum Distance function
}
\examples{
\dontrun{
l1 <- c(182, 180, 178, 176)
l2 <- c(178, 180, 178, 180)
dis <- dist_euclidian(l1, l2)
}
}
