% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{dist_num_mismatches}
\alias{dist_num_mismatches}
\title{An Number of Non-matches Distance function}
\usage{
dist_num_mismatches(multilocus1, multilocus2)
}
\arguments{
\item{multilocus1}{A vector with the locuses of the first object to be compared.}

\item{multilocus2}{A vector with the locuses of the second object to be compared.}
}
\value{
The number of locuses that did not match. Missing data are counted as matches.
}
\description{
An Number of Non-matches Distance function
}
\examples{
\dontrun{
l1 <- c(182, 180, 178, 176)
l2 <- c(178, 180, 178, 178)
dis <- dist_euclidian(l1, l2)
}
}
