% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{extract_one_index_from_batch}
\alias{extract_one_index_from_batch}
\title{Extract One Index From Batch of New_Data}
\usage{
extract_one_index_from_batch(batch, index)
}
\arguments{
\item{batch}{A new_data object.}

\item{index}{The index of the new_data to be extracted.}
}
\value{
A new_data object that only contains one point of new data.
}
\description{
Extract One Index From Batch of New_Data
}
\examples{
\dontrun{
new_data_one <- extract_one_index_from_batch(batch = new_data, index = "SEP0159539")
}
}
