% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{generate_allelemtach_profile_plot}
\alias{generate_allelemtach_profile_plot}
\title{Generate Allelematch Profile Plot}
\usage{
generate_allelemtach_profile_plot(am_dataset)
}
\arguments{
\item{am_dataset}{A allelematch dataset to examine the optimal mismatch value for}
}
\value{
The plotdata to show the user what the program thinks is the optimal mismatch value
}
\description{
Generate Allelematch Profile Plot
}
\examples{
\dontrun{
library(shiny)

output$plot <- renderPlot({
  generate_allelematch_plot(am_data)
})
}
}
