% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{generate_threshold_plot}
\alias{generate_threshold_plot}
\title{Generate threshold plot}
\usage{
generate_threshold_plot(new_data, data)
}
\arguments{
\item{new_data}{A new_data object.}

\item{data}{A data object.}
}
\value{
Renders a plot showing how different thresholds would affect the resulting merge.
importFrom graphics legend lines
}
\description{
Generate threshold plot
}
\examples{
\dontrun{
generate_threshold_plot(new_data = new_data, data = data)
}
}
