% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_id}
\alias{get_id}
\title{Simplify the ID getting process}
\usage{
get_id(row)
}
\arguments{
\item{row}{The row for which the id wishes to be taken}
}
\value{
The relevant id
}
\description{
Return the override id if there is one, otherwise return the individ_id
}
\examples{
\dontrun{
# Get the id for the sample at the 17th row
id <- get_id(search_data[[17,]])
}
}
