% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{handle_multimatch}
\alias{handle_multimatch}
\title{Handle Multimatch By User}
\usage{
handle_multimatch(search_data, multiple_matches, multimatch_index, new_id)
}
\arguments{
\item{search_data}{The search data, a dateframe with indexes, locuses and ids}

\item{multiple_matches}{A list with the indexes that currently matches to multiples ids in the search_data}

\item{multimatch_index}{The index of the sample that is to get a specified id}

\item{new_id}{The new id that the specified id and all entries in the same group will get}
}
\value{
A list with the updated search_data and multiple_matches
}
\description{
This function takes
}
\examples{
\dontrun{
c(search_data, multiple_match) \%<-\% handle_multimatch
           (search_data, multiple_matches, "SEP123", "B31")
}
}
