% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{load_data}
\alias{load_data}
\title{Load Original Dataset}
\usage{
load_data(
  file_path,
  index_column,
  locus_columns,
  individ_column = NA,
  meta_columns,
  na_strings = c("NA", "-99", "0", "000"),
  sheet = 1
)
}
\arguments{
\item{file_path}{The path to a file, either a .csv, .ods, excel file or SQLite database.}

\item{index_column}{The name of the column that contains indexes, a unique identifier for each row of data. In the example data
this column is called "index"}

\item{locus_columns}{A vector with the columns that contain the locus data. In the example data these are: c("G10L_1", "G10L_2",
"MU05_1", "MU05_2", ..., "MU59_2").}

\item{individ_column}{The name of the column that contains the id of the individual the sample belongs to. When loading new data to
be compared this is set to NA as the new data don't have any individual-classification}

\item{meta_columns}{A vector with the columns that contain the meta-data for each row. In the example data these are: c(date = "date",
north = "north", east = "east", gender = "gender"). The vector's names are required to be date, north, east, and gender.}

\item{na_strings}{A vector of strings that are to be interpreted as NA values. Default is "NA", "-99", "0", and "000".}

\item{sheet}{If a .ods or excel file is loaded, this variable is required and specifies the sheet to be loaded.}
}
\value{
A data object that is used in the rest of this package.
}
\description{
Load Original Dataset
}
\examples{
\dontrun{
See workflow.R
}
}
