% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Helper Function to load different forms of data}
\usage{
load_raw_data(file_path, sheet = 1, na_strings = c("NA", "-99", "0", "000"))
}
\arguments{
\item{file_path}{The path to a file, either a .csv, .ods, excel file or SQLite database.}

\item{sheet}{If a .ods or excel file is loaded, this variable is required and specifies the sheet to be loaded.}

\item{na_strings}{A vector of strings that are to be interpreted as NA values. Default is "NA", "-99", "0", and "000".}
}
\value{
A table with all columns of the file or database
}
\description{
Helper Function to load different forms of data
}
\examples{
\dontrun{
raw_data <- load_raw_data(file_path = "./data/bear_data_all.csv")
}
}
