% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{match_new_data}
\alias{match_new_data}
\title{Title}
\usage{
match_new_data(data, new_data, additional_data_columns, allele_mismatch)
}
\arguments{
\item{data}{The old data, containing all samples that have already been handeled, or been choosen not to be handleh}

\item{new_data}{The new data, the meta-data for the new sample, containing the same columns as the old data, index, locus, date, gender, etc}

\item{additional_data_columns}{The names of the columns containing the meta-data for the \code{\link{create_allelematch_dataset}} to be able to create a new am_dataset}

\item{allele_mismatch}{The mismatch (numbers of alleles that are allowed to differ between individuals) to use for this grouping, user could be intrested in being more strict or more loose, up to them}
}
\value{
Returns only the new data grouped into search_data format, which of them that matched multiple and if some of them were unclassified
}
\description{
Title
}
