% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{merge_new_data}
\alias{merge_new_data}
\title{Merge New Data Into The Dataset}
\usage{
merge_new_data(new_data, data, new_data_id)
}
\arguments{
\item{new_data}{A new_data object with one datapoint. Generated by extract_one_index_from_batch.}

\item{data}{A data object.}

\item{new_data_id}{The new individ-id for the datapoint in new_data.}
}
\value{
A list with $data that contains the new data and $success that is either TRUE or FALSE depending on if the function succeeded.
The test is rudimentary and weird/broken or corrupt data may be generated even if this value is TRUE.
}
\description{
Merge New Data Into The Dataset
}
\examples{
\dontrun{
ind <- "SEP0159539"
new_id <- possible_matches[[ind]]$ids[1]
merged_data <- merge_new_data(new_data = extract_one_index_from_batch(batch = new_data,
    index = ind), data = data, new_data_id = new_id)
}
}
