% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{sanity_check_new_data}
\alias{sanity_check_new_data}
\title{Sanity Check New Data}
\usage{
sanity_check_new_data(new_data, data)
}
\arguments{
\item{new_data}{A new_data object, created by the "create_new_data_batch" and "create_new_data" functions.}

\item{data}{A data object, created by the "load_data" function.}
}
\value{
A vector of messaged that contains information on the sanity of the new data.
}
\description{
Sanity Check New Data
}
\examples{
\dontrun{
sanity_message <- sanity_check_new_data(new_data = new_data, data = data)
print(sanity_message)
}
}
