% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow_example}
\alias{workflow_example}
\title{Workflow example}
\usage{
workflow_example()
}
\description{
Do not run this function, copy it to a file and configure it to your environment. Use the test data bear_data_part1.csv, bear_data_part2.csv and bear_data_all.csv or create your own data. Information on constraints and configuration of new data exists on \verb{https://github.com/CGI-NRM/GenotypeCheck}.
}
