% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{combine_multilocus}
\alias{combine_multilocus}
\title{Title}
\usage{
combine_multilocus(locus)
}
\arguments{
\item{locus}{A vector with all the locus in string format.}
}
\value{
A long string with all locus combined, padding zeros making ever locus three characters long.
}
\description{
Title
}
\examples{
\dontrun{
multilocus <- c("182", "180", "152", "152")
multilocus_combined <- combine_multilocus(multilocus)
}
\dontrun{
multilocus_combined <- apply(multilocus_matrix, 1, combine_multilocus)
}

}
