% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{create_new_data_batch}
\alias{create_new_data_batch}
\title{Load A File Containing New Data}
\usage{
create_new_data_batch(raw_data, index_column, locus_columns, meta_columns)
}
\arguments{
\item{raw_data}{A table/data.frame with all columns that are to be extracted. Generated by the "load_raw_data" function.}

\item{index_column}{The name of the column containing the index.}

\item{locus_columns}{A vector with the names of the columns that contains the locus data.}

\item{meta_columns}{A vector with the names of the columns that contains the meta data.}
}
\value{
A new_data object, similar to a data object but missing the "individ" column.
}
\description{
This is a wrapper function for load_data function.
}
\examples{
\dontrun{
See "?load_data" and "?create_new_data"
}
}
