% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{create_search_data}
\alias{create_search_data}
\title{Load Data and Group Into Individuals}
\usage{
create_search_data(data, am_data, allele_mismatch)
}
\arguments{
\item{data}{A dateframe with index, all meta-data and locus}

\item{am_data}{An allelematch dataset, converted from the data containing the index and locus}

\item{allele_mismatch}{A value for how many allele mismatchs are to be allowed and still count like a match}
}
\value{
A list with the search_data, which is a list with $index, $multilocus, and $individ_id, a list with the $index of the samples that matched multiple individuals, and a list with the $index and $multilocus of the samples were unclassified.
}
\description{
TODO: Keep the multimatch data
}
\examples{
\dontrun{
search_data <- create_search_data("data.csv")
}
}
