% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{generate_user_choice_data_frame}
\alias{generate_user_choice_data_frame}
\title{Gather all of the data the user could be interested in when choosing to merge the new_data.}
\usage{
generate_user_choice_data_frame(possible_matches, new_data, data, ind)
}
\arguments{
\item{possible_matches}{A possible_matches object generated by the "match_new_data" function}

\item{new_data}{A new_data object.}

\item{data}{A data object.}

\item{ind}{The index for which the dataframe should be generated.}
}
\value{
A dataframe with a index, multilocus, distance, and individ column. All data of every individual where atleast one point was
within the threshold is included in the dataframe.
}
\description{
Gather all of the data the user could be interested in when choosing to merge the new_data.
}
\examples{
\dontrun{
df <- generate_user_choice_data_frame(possible_matches = possible_matches,
    new_data = new_data, data = data, ind = "SEP0159539")
}
}
