% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import and format data}
\usage{
import_data(file)
}
\arguments{
\item{file}{The path to the file to be imported}
}
\value{
A table with the relevant data
}
\description{
Import and format data
}
\examples{
data <- import_data("raw_data.csv")
}
