% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{load_data}
\alias{load_data}
\title{Load Original Dataset}
\usage{
load_data(
  raw_data,
  index_column,
  locus_columns,
  individ_column = NA,
  meta_columns
)
}
\arguments{
\item{raw_data}{A table/data.frame with all columns that are to be extracted. Generated by the "load_raw_data" function.}

\item{index_column}{The name of the column that contains indexes, a unique identifier for each row of data. In the example data
this column is called "index"}

\item{locus_columns}{A vector with the columns that contain the locus data. In the example data these are: c("G10L_1", "G10L_2",
"MU05_1", "MU05_2", ..., "MU59_2").}

\item{individ_column}{The name of the column that contains the id of the individual the sample belongs to. When loading new data to
be compared this is set to NA as the new data don't have any individual-classification}

\item{meta_columns}{A vector with the columns that contain the meta-data for each row. In the example data these are: c(date = "date",
north = "north", east = "east", gender = "gender"). The vector's names are required to be date, north, east, and gender.}
}
\value{
A list containing two object. 1. Data object that is used in the rest of this package. 2. A success message which contains usefull information.
}
\description{
Load Original Dataset
}
\examples{
\dontrun{
locus_columns <- c("MU09_1", "MU09_2", "MU10_1", "MU10_2", "MU05_1", "MU05_2",
    "MU23_1", "MU23_2", "MU51_1", "MU51_2", "MU59_1", "MU59_2", "G10L_1",
    "G10L_2", "MU50_1", "MU50_2")
names(locus_columns) <- c("MU09 - 1", "MU09 - 2", "MU10 - 1", "MU10 - 2", "MU05 - 1", "MU05 - 2",
    "MU23 - 1", "MU23 - 2", "MU51 - 1", "MU51 - 2", "MU59 - 1", "MU59 - 2", "G10L - 1",
    "G10L - 2", "MU50 - 1", "MU50 - 2")
c(data, success_message) \%<-\% load_data(load_raw_data(file_path = "./data/bear_data_all.csv"), index_column = "index",
    locus_columns = locus_columns, individ_column = "individ", meta_columns = c(gender = "gender",
}
}
