% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{load_data_sqlite}
\alias{load_data_sqlite}
\title{(load_data helper function) Loads a SQLite database from a file}
\usage{
load_data_sqlite(file_path, table = "Bears")
}
\arguments{
\item{file_path}{The path to the *.db file to be loaded.}

\item{table}{The name of the table in the database that contains the desired data. Default is "Bears".}
}
\value{
A table with all data from the table
}
\description{
(load_data helper function) Loads a SQLite database from a file
}
\examples{
\dontrun{
raw_data <- load_data_sqlite(file_path = "path/to/file.db", table = "Bears")
}
}
