% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{match_new_data}
\alias{match_new_data}
\title{Match New Data (Generate possible_matches object)}
\usage{
match_new_data(new_data, threshold)
}
\arguments{
\item{new_data}{A new_data object.}

\item{threshold}{A value for under which threshold the new data will be matched to the existing data.}
}
\value{
A named list containing a vector of all data with a distance under the threshold. If the new_data$distances is generated as
expected, new_data can be matched against both new_data and data points. Be aware the refrenced indexes might exist in different objects.
}
\description{
Match New Data (Generate possible_matches object)
}
\examples{
\dontrun{
possible_matches <- match_new_data(new_data = new_data, threshold = threshold)
}
}
